Cluster magic.

+++ Trunk/src/hearthstone-world/WorldSocket.h
@@ -121,7 +121,7 @@ public:
 	HEARTHSTONE_INLINE uint32 GetRemotePort() { return ntohs(m_address.sin_port); }
 
 	HEARTHSTONE_INLINE void SendPacket(WorldPacket* packet) { if(!packet) return; OutPacket(packet->GetOpcode(), (uint16)packet->size(), (packet->size() ? (const void*)packet->contents() : NULL)); }
-	HEARTHSTONE_INLINE void SendPacket(StackPacket * packet) { if(!packet) return; OutPacket(packet->GetOpcode(), packet->GetSize(), (packet->GetSize() ? (const void*)packet->GetBufferPointer() : NULL)); }
+	HEARTHSTONE_INLINE void SendPacket(StackPacket * packet) { if(!packet) return; OutPacket(packet->GetOpcode(), (uint16)packet->GetSize(), (packet->GetSize() ? (const void*)packet->GetBufferPointer() : NULL)); }
 	void __fastcall OutPacket(uint16 opcode, uint16 len, const void* data);
 	HEARTHSTONE_INLINE uint32 GetSessionId() { return m_sessionId; }
